+++ rimopy/eli.py
@@ -28,7 +28,7 @@ It exports the following functions:
 
 from dataclasses import dataclass
 import math
-from typing import List, Union
+from typing import List, Union, Iterable
 
 from . import spice_iface, esi, elref
 from .types import MoonData
@@ -52,10 +52,14 @@ class ELISettings:
         If True the ROLO model reflectance will be adjusted using Apollo spectra, in case it's
         calculated interpolating surrounding reflectances. The Apollo spectra is the spectra
         generated with the Moon samples from Apollo 16th mission.
+    per_nm : bool
+        If True the ELI will be in Wm⁻²/nm, otherwise it will be in Wm⁻². Default is False.
     """
     apply_correction: bool = False
     interpolate_rolo_coefficients: bool = False
     adjust_apollo: bool = True
+    per_nm: bool = False
+
 
 @dataclass
 class EarthPoint:
@@ -111,7 +115,7 @@ class EarthPoint:
             self.utc_times = [utc_times]
 
 
-def _get_esi(esi_calc: esi.ESICalculator, wavelength_nm: float) -> float:
+def _get_esi(esi_calc: esi.ESICalculator, wavelength_nm: float, eli_settings: ELISettings) -> float:
     """Gets the expected extraterrestrial solar irradiance at a concrete wavelength
     Returns the data in Wm⁻²
 
@@ -123,36 +127,19 @@ def _get_esi(esi_calc: esi.ESICalculator, wavelength_nm: float) -> float:
     wavelength_nm : float
         Wavelength (in nanometers) of which the extraterrestrial solar irradiance will be
         obtained
+    eli_settings : ELISettings
+        Configuration of the ELI calculation method.
 
     Returns
     -------
     float
-        The expected extraterrestrial solar irradiance in W/sm
+        The expected extraterrestrial solar irradiance in Wm⁻² or Wm⁻²/nm
     """
-    return esi_calc.get_esi(wavelength_nm)
+    return esi_calc.get_esi(wavelength_nm, eli_settings.per_nm)
 
-def _get_esi_per_nm(esi_calc: esi.ESICalculator, wavelength_nm: float) -> float:
-    """Gets the expected extraterrestrial solar irradiance at a concrete wavelength
-    Returns the data in Wm⁻²/nm
-
-    Parameters
-    ----------
-    esi_calc : esi.ESICalculator
-        ESI Calculator that will be used in the calculation of the Extraterrestrial Solar
-        Irradiance.
-    wavelength_nm : float
-        Wavelength (in nanometers) of which the extraterrestrial solar irradiance will be
-        obtained
-
-    Returns
-    -------
-    float
-        The expected extraterrestrial solar irradiance in W/sm
-        """
-    return esi_calc.get_esi_per_nm(wavelength_nm)
 
 def _calculate_eli(wavelength_nm: float, moon_data: MoonData, esi_calc: esi.ESICalculator,
-                   eli_settings: ELISettings, per_nm: bool = False) -> float:
+                   eli_settings: ELISettings) -> float:
     """Calculation of Extraterrestrial Lunar Irradiance following Eq 3 in Roman et al., 2020
 
     Simulates a lunar observation for a wavelength for any observer/solar selenographic
@@ -170,8 +157,6 @@ def _calculate_eli(wavelength_nm: float, moon_data: MoonData, esi_calc: esi.ESIC
         Irradiance.
     eli_settings : ELISettings
         Configuration of the ELI calculation method.
-    per_nm : bool
-        True if the user wants the ELI in Wm⁻²/nm, otherwise it will be in Wm⁻². Default is False.
 
     Returns
     -------
@@ -196,10 +181,7 @@ def _calculate_eli(wavelength_nm: float, moon_data: MoonData, esi_calc: esi.ESIC
 
     solid_angle_moon: float = 6.4177e-05
     omega = solid_angle_moon
-    if per_nm:
-        esk = _get_esi_per_nm(esi_calc, wavelength_nm)
-    else:
-        esk = _get_esi(esi_calc, wavelength_nm)
+    esk = _get_esi(esi_calc, wavelength_nm, eli_settings)
     dsm = moon_data.distance_sun_moon
     dom = moon_data.distance_observer_moon
     distance_earth_moon_km: int = 384400
@@ -208,9 +190,10 @@ def _calculate_eli(wavelength_nm: float, moon_data: MoonData, esi_calc: esi.ESIC
         (distance_earth_moon_km / dom) ** 2
     return lunar_irr
 
-def get_eli_bypass(wavelength_nm: Union[float, List[float]], moon_data: MoonData,
+
+def get_eli_bypass(wavelength_nm: Iterable[float], moon_data: MoonData,
                    esi_calc: esi.ESICalculator = None,
-                   eli_settings: ELISettings = None) -> Union[float, List[float]]:
+                   eli_settings: ELISettings = None) -> List[float]:
     """Calculation of Extraterrestrial Lunar Irradiance following Eq 3 in Roman et al., 2020
 
     Allow users to simulate lunar observation for any observer/solar selenographic
@@ -221,8 +204,8 @@ def get_eli_bypass(wavelength_nm: Union[float, List[float]], moon_data: MoonData
 
     Parameters
     ----------
-    wavelength_nm : float | list of float
-        Wavelength/s (in nanometers) of which the extraterrestrial lunar irradiance will be
+    wavelength_nm : iterable of float
+        Wavelengths (in nanometers) of which the extraterrestrial lunar irradiance will be
         calculated.
     moon_data : MoonData
         Moon data needed to calculate Moon's irradiance
@@ -234,27 +217,25 @@ def get_eli_bypass(wavelength_nm: Union[float, List[float]], moon_data: MoonData
 
     Returns
     -------
-    float | list of float
-        The extraterrestrial lunar irradiance/s calculated. It will be a list if parameter
-        "wavelength_nm" was a list.
+    list of float
+        The extraterrestrial lunar irradiances calculated.
     """
     if eli_settings is None:
         eli_settings = ELISettings()
     if esi_calc is None:
         esi_calc = esi.ESICalculator()
-    if isinstance(wavelength_nm, list):
-        elis = []
-        for wlen in wavelength_nm:
-            elis.append(_calculate_eli(wlen, moon_data, esi_calc, eli_settings))
-        return elis
-    return _calculate_eli(wavelength_nm, moon_data, esi_calc, eli_settings)
+    elis = []
+    for wlen in wavelength_nm:
+        elis.append(_calculate_eli(wlen, moon_data, esi_calc, eli_settings))
+    return elis
 
-def get_eli_from_extra_kernels(wavelength_nm: Union[float, List[float]],
+
+def get_eli_from_extra_kernels(wavelength_nm: Iterable[float],
         utc_times: Union[str, List[str]], kernels_path: str, extra_kernels: List[str],
         extra_kernels_path: str, observer_name: str,
         esi_calc: esi.ESICalculator = esi.ESICalculator(),
         eli_settings: ELISettings = ELISettings()
-        ) -> Union[float, List[float], List[List[float]]]:
+        ) -> Union[List[float], List[List[float]]]:
     """Calculation of Extraterrestrial Lunar Irradiance from geographic coordinates
 
     Allow users to simulate lunar observations for any observer position around the Earth
@@ -267,8 +248,8 @@ def get_eli_from_extra_kernels(wavelength_nm: Union[float, List[float]],
 
     Parameters
     ----------
-    wavelength_nm : float | list of float
-        Wavelength/s (in nanometers) of which the extraterrestrial lunar irradiance will be
+    wavelength_nm : iterable of float
+        Wavelengths (in nanometers) of which the extraterrestrial lunar irradiance will be
         calculated.
     utc_times: str | list of str
         Time/s at which the ELI will be calculated, in a valid UTC DateTime format.
@@ -288,11 +269,11 @@ def get_eli_from_extra_kernels(wavelength_nm: Union[float, List[float]],
 
     Returns
     -------
-    float | list of float | list of list of float
-        The extraterrestrial lunar irradiance/s calculated. It will be a list if parameter
-        "wavelength_nm" was a list, or if EarthPoint utc_times is a list. In case both are lists,
-        it will be a list of lists of floats. The list will contain a list for every utc_time
-        present, and each of those will contain all irradiances associated for each wavelength.
+    list of float | list of list of float
+        The extraterrestrial lunar irradiances calculated. If EarthPoint utc_times is a list
+        it will be a list of lists of floats, otherwise a simple list. The list will contain
+        a list for every utc_time present, and each of those will contain all irradiances
+        associated for each wavelength.
     """
     moon_data = spice_iface.get_moon_datas_from_extra_kernels(utc_times, kernels_path,
                                                               extra_kernels, extra_kernels_path,
@@ -304,10 +285,10 @@ def get_eli_from_extra_kernels(wavelength_nm: Union[float, List[float]],
         return irradiances[0]
     return irradiances
 
-def get_eli(wavelength_nm: Union[float, List[float]], earth_data: EarthPoint, kernels_path: str,
+def get_eli(wavelength_nm: Iterable[float], earth_data: EarthPoint, kernels_path: str,
             esi_calc: esi.ESICalculator = None,
             eli_settings: ELISettings = None,
-            ) -> Union[float, List[float], List[List[float]]]:
+            ) -> Union[List[float], List[List[float]]]:
     """Calculation of Extraterrestrial Lunar Irradiance from geographic coordinates
 
     Allow users to simulate lunar observations for any observer position around the Earth
@@ -317,8 +298,8 @@ def get_eli(wavelength_nm: Union[float, List[float]], earth_data: EarthPoint, ke
 
     Parameters
     ----------
-    wavelength_nm : float | list of float
-        Wavelength/s (in nanometers) of which the extraterrestrial lunar irradiance will be
+    wavelength_nm : iterable of float
+        Wavelengths (in nanometers) of which the extraterrestrial lunar irradiance will be
         calculated.
     earth_data : EarthPoint
         Data of the point on Earth surface of which the ELI will be calculated.
@@ -332,158 +313,17 @@ def get_eli(wavelength_nm: Union[float, List[float]], earth_data: EarthPoint, ke
 
     Returns
     -------
-    float | list of float | list of list of float
-        The extraterrestrial lunar irradiance/s calculated. It will be a list if parameter
-        "wavelength_nm" was a list, or if EarthPoint utc_times is a list. In case both are lists,
-        it will be a list of lists of floats. The list will contain a list for every utc_time
-        present, and each of those will contain all irradiances associated for each wavelength.
+    list of float | list of list of float
+        The extraterrestrial lunar irradiances calculated. If EarthPoint utc_times is a list
+        it will be a list of lists of floats, otherwise a simple list. The list will contain
+        a list for every utc_time present, and each of those will contain all irradiances
+        associated for each wavelength.
     """
     moon_data = spice_iface.get_moon_datas(earth_data.lat, earth_data.lon, earth_data.altitude,
                                            earth_data.utc_times, kernels_path)
     irradiances = []
     for moon_d in moon_data:
         irradiances.append(get_eli_bypass(wavelength_nm, moon_d, esi_calc, eli_settings))
-    if len(irradiances) == 1:
-        return irradiances[0]
-    return irradiances
-
-def get_eli_bypass_per_nm(wavelength_nm: Union[float, List[float]], moon_data: MoonData,
-                          esi_calc: esi.ESICalculator = None,
-                          eli_settings: ELISettings = None,
-                          ) -> Union[float, List[float]]:
-    """Calculation of Extraterrestrial Lunar Irradiance following Eq 3 in Roman et al., 2020
-
-    Allow users to simulate lunar observation for any observer/solar selenographic
-    latitude and longitude (thus bypassing the need for their computation from the
-    position/time of the observer).
-
-    Returns the data in Wm⁻²/nm
-
-    Parameters
-    ----------
-    wavelength_nm : float | list of float
-        Wavelength/s (in nanometers) of which the extraterrestrial lunar irradiance will be
-        calculated.
-    moon_data : MoonData
-        Moon data needed to calculate Moon's irradiance.
-    esi_calc : esi.ESICalculator
-        ESI Calculator that will be used in the calculation of the Extraterrestrial Solar
-        Irradiance.
-    eli_settings : ELISettings
-        Configuration of the ELI calculation method.
-
-    Returns
-    -------
-    float | list of float
-        The extraterrestrial lunar irradiance/s calculated. It will be a list if parameter
-        "wavelength_nm" was a list.
-    """
-    if eli_settings is None:
-        eli_settings = ELISettings()
-    if esi_calc is None:
-        esi_calc = esi.ESICalculator()
-    if isinstance(wavelength_nm, list):
-        elis = []
-        for wlen in wavelength_nm:
-            elis.append(_calculate_eli(wlen, moon_data, esi_calc, eli_settings, True))
-        return elis
-    return _calculate_eli(wavelength_nm, moon_data, esi_calc, eli_settings, True)
-
-def get_eli_per_nm_from_extra_kernels(wavelength_nm: Union[float, List[float]],
-        utc_times: Union[str, List[str]], kernels_path: str, extra_kernels: List[str],
-        extra_kernels_path: str, observer_name: str,
-        esi_calc: esi.ESICalculator = None,
-        eli_settings: ELISettings = None,
-        ) -> Union[float, List[float], List[List[float]]]:
-    """Calculation of Extraterrestrial Lunar Irradiance from geographic coordinates
-
-    Allow users to simulate lunar observations for any observer position around the Earth
-    and at any time.
-
-    It loads the observer body data from custom extra kernels instead of generating it from
-    basic kernels.
-
-    Returns the data in Wm⁻²/nm
-
-    Parameters
-    ----------
-    wavelength_nm : float | list of float
-        Wavelength/s (in nanometers) of which the extraterrestrial lunar irradiance will be
-        calculated.
-    utc_times: str | list of str
-        Time/s at which the ELI will be calculated, in a valid UTC DateTime format.
-    kernels_path : str
-        Folder where the needed SPICE kernels are stored.
-    extra_kernels: list of str
-        Custom kernels from which the observer body will be loaded, instead of calculating it.
-    extra_kernels_path: str
-        Folder where the extra kernels are located.
-    observer_name: str
-        Name of the body of the observer that will be loaded from the extra kernels.
-    esi_calc : esi.ESICalculator
-        ESI Calculator that will be used in the calculation of the Extraterrestrial Solar
-        Irradiance.
-    eli_settings : ELISettings
-        Configuration of the ELI calculation method.
-
-    Returns
-    -------
-    float | list of float | list of list of float
-        The extraterrestrial lunar irradiance/s calculated. It will be a list if parameter
-        "wavelength_nm" was a list, or if EarthPoint utc_times is a list. In case both are lists,
-        it will be a list of lists of floats. The list will contain a list for every utc_time
-        present, and each of those will contain all irradiances associated for each wavelength.
-    """
-    moon_data = spice_iface.get_moon_datas_from_extra_kernels(utc_times, kernels_path,
-                                                              extra_kernels, extra_kernels_path,
-                                                              observer_name)
-    irradiances = []
-    for moon_d in moon_data:
-        irradiances.append(get_eli_bypass_per_nm(wavelength_nm, moon_d, esi_calc, eli_settings))
-    if len(irradiances) == 1:
-        return irradiances[0]
-    return irradiances
-
-def get_eli_per_nm(wavelength_nm: Union[float, List[float]], earth_data: EarthPoint,
-                   kernels_path: str, esi_calc: esi.ESICalculator = None,
-                   eli_settings: ELISettings = None,
-                   ) -> Union[float, List[float], List[List[float]]]:
-    """Calculation of Extraterrestrial Lunar Irradiance from geographic coordinates
-
-    Allow users to simulate lunar observations for any observer position around the Earth
-    and at any time.
-
-    Returns the data in Wm⁻²/nm
-
-    Parameters
-    ----------
-    wavelength_nm : float | list of float
-        Wavelength/s (in nanometers) of which the extraterrestrial lunar irradiance will be
-        calculated.
-    earth_data : EarthPoint
-        Data of the point on Earth surface of which the ELI will be calculated.
-    kernels_path : str
-        Folder where the needed SPICE kernels are stored.
-    esi_calc : esi.ESICalculator
-        ESI Calculator that will be used in the calculation of the Extraterrestrial Solar
-        Irradiance.
-    eli_settings : ELISettings
-        Configuration of the ELI calculation method.
-
-    Returns
-    -------
-    float | list of float | list of list of float
-        The extraterrestrial lunar irradiance/s calculated. It will be a list if parameter
-        "wavelength_nm" was a list, or if EarthPoint utc_times is a list. In case both are lists,
-        it will be a list of lists of floats. The list will contain a list for every utc_time
-        present, and each of those will contain all irradiances associated for each wavelength.
-    """
-    moon_data = spice_iface.get_moon_datas(earth_data.lat, earth_data.lon,
-                                               earth_data.altitude, earth_data.utc_times,
-                                               kernels_path)
-    irradiances = []
-    for moon_d in moon_data:
-        irradiances.append(get_eli_bypass_per_nm(wavelength_nm, moon_d, esi_calc, eli_settings))
     if len(irradiances) == 1:
         return irradiances[0]
     return irradiances